app/components/error-boundary/error-boundary.tsx
import { type ErrorResponse } from "@remix-run/node";
import {
  isRouteErrorResponse,
  useParams,
  useRouteError,
} from "@remix-run/react";

export function getErrorMessage(error: unknown) {
  if (typeof error === "string") return error;
  if (
    error &&
    typeof error === "object" &&
    "message" in error &&
    typeof error.message === "string"
  ) {
    return error.message;
  }
  console.error("Unable to get error message for error", error);
  return "Unknown Error";
}

type StatusHandler = (info: {
  error: ErrorResponse;
  params: Record<string, string | undefined>;
}) => JSX.Element | null;

type ErrorBoundaryComponentOptions = {
  defaultStatusHandler?: StatusHandler;
  statusHandlers?: Record<number, StatusHandler>;
  unexpectedErrorHandler?: (error: unknown) => JSX.Element | null;
};
export function ErrorBoundaryComponent({
  defaultStatusHandler = ({ error }) => (
    <div>
      <p>
        {error.status} {error.data}
      </p>
    </div>
  ),
  statusHandlers,
  unexpectedErrorHandler = (error) => <p>{getErrorMessage(error)}</p>,
}: ErrorBoundaryComponentOptions) {
  const error = useRouteError();
  const params = useParams();

  if (typeof document !== "undefined") {
    console.error(error);
  }

  return (
    <div className="container mx-auto flex h-full w-full items-center justify-center p-20 text-xl">
      {isRouteErrorResponse(error)
        ? (statusHandlers?.[error.status] ?? defaultStatusHandler)({
            error,
            params,
          })
        : unexpectedErrorHandler(error)}
    </div>
  );
}
